post module completed

+++ app/Http/Controllers/AdminUsersController.php
@@ -74,7 +74,6 @@ class AdminUsersController extends Controller
 
         }
 
-        $input['password'] = bcrypt($request->password);
 
         User::create($input);
 
@@ -165,7 +164,10 @@ class AdminUsersController extends Controller
         //
         $user = User::findOrFail($id);
 
-        unlink(public_path().$user->photo->file);
+        if($user->photo_id) {
+
+            unlink(public_path() . $user->photo->file);
+        }
 
         $user->delete();
